Reorganize the crate to better accommodate sinks
- Moves most futures-related items into a futures submodule, with
  deprecated re-exports.

- Splits task into two modules: task and executor.
