+++ src/future/and_then.rs
@@ -1,5 +1,5 @@
 use {Future, IntoFuture, Poll};
-use chain::Chain;
+use super::chain::Chain;
 
 /// Future for the `and_then` combinator, chaining a computation onto the end of
 /// another future which completes successfully.
