+++ src/future/collect.rs
