+++ src/future/flatten_stream.rs
