+++ src/future/fuse.rs
