+++ src/future/lazy.rs
