+++ src/future/map_err.rs
