+++ src/future/then.rs
@@ -1,5 +1,5 @@
 use {Future, IntoFuture, Poll};
-use chain::Chain;
+use super::chain::Chain;
 
 /// Future for the `then` combinator, chaining computations on the end of
 /// another future regardless of its outcome.
