+++ src/task_impl/mod.rs
@@ -1,32 +1,3 @@
-//! Tasks used to drive a future computation
-//!
-//! It's intended over time a particular operation (such as servicing an HTTP
-//! request) will involve many futures. This entire operation, however, can be
-//! thought of as one unit, as the entire result is essentially just moving
-//! through one large state machine.
-//!
-//! A "task" is the unit of abstraction for what is driving this state machine
-//! and tree of futures forward. A task is used to poll futures and schedule
-//! futures with, and has utilities for sharing data between tasks and handles
-//! for notifying when a future is ready. Each task also has its own set of
-//! task-local data generated by `task_local!`.
-//!
-//! Note that libraries typically should not manage tasks themselves, but rather
-//! leave that to event loops and other "executors", or by using the `wait`
-//! method to create and execute a task directly on the current thread.
-//!
-//! There are two basic execution models for tasks: manually with a `Notify`
-//! implementation or via an `Executor`. A `Notify` is just a source of
-//! notifications that a future is ready to progress, where an `Executor` is
-//! generally one or more threads together with a queue of tasks to execute.
-//! Again, though, normally these methods are not called manually.
-//!
-//! ## Functions
-//!
-//! There is an important bare function in this module: `park`. The `park`
-//! function is similar to the standard library's `thread::park` method where it
-//! returns a handle to wake up a task at a later date (via an `unpark` method).
-
 use std::prelude::v1::*;
 
 use std::cell::Cell;
@@ -37,9 +8,10 @@ use std::thread;
 
 use {BoxFuture, Poll, Future, Async};
 use stream::Stream;
-use task::unpark_mutex::UnparkMutex;
 
 mod unpark_mutex;
+use self::unpark_mutex::UnparkMutex;
+
 mod task_rc;
 mod data;
 pub use self::task_rc::TaskRc;
@@ -163,6 +135,97 @@ impl fmt::Debug for Task {
         f.debug_struct("Task")
          .field("id", &self.id)
          .finish()
+    }
+}
+
+/// For the duration of the given callback, add an "unpark event" to be
+/// triggered when the task handle is used to unpark the task.
+///
+/// Unpark events are used to pass information about what event caused a task to
+/// be unparked. In some cases, tasks are waiting on a large number of possible
+/// events, and need precise information about the wakeup to avoid extraneous
+/// polling.
+///
+/// Every `Task` handle comes with a set of unpark events which will fire when
+/// `unpark` is called. When fired, these events insert an identifer into a
+/// concurrent set, which the task can read from to determine what events
+/// occurred.
+///
+/// This function immediately invokes the closure, `f`, but arranges things so
+/// that `task::park` will produce a `Task` handle that includes the given
+/// unpark event.
+///
+/// # Panics
+///
+/// This function will panic if a task is not currently being executed. That
+/// is, this method can be dangerous to call outside of an implementation of
+/// `poll`.
+pub fn with_unpark_event<F, R>(event: UnparkEvent, f: F) -> R
+    where F: FnOnce() -> R
+{
+    with(|task, data| {
+        let new_task = Task {
+            id: task.id,
+            unpark: task.unpark.clone(),
+            events: task.events.with_event(event),
+        };
+        set(&new_task, data, f)
+    })
+}
+
+#[derive(Clone)]
+/// A set insertion to trigger upon `unpark`.
+///
+/// Unpark events are used to communicate information about *why* an unpark
+/// occured, in particular populating sets with event identifiers so that the
+/// unparked task can avoid extraneous polling. See `with_unpark_event` for
+/// more.
+pub struct UnparkEvent {
+    set: Arc<EventSet>,
+    item: usize,
+}
+
+impl UnparkEvent {
+    /// Construct an unpark event that will insert `id` into `set` when
+    /// triggered.
+    pub fn new(set: Arc<EventSet>, id: usize) -> UnparkEvent {
+        UnparkEvent {
+            set: set,
+            item: id,
+        }
+    }
+}
+
+/// A concurrent set which allows for the insertion of `usize` values.
+///
+/// `EventSet`s are used to communicate precise information about the event(s)
+/// that trigged a task notification. See `task::with_unpark_event` for details.
+pub trait EventSet: Send + Sync + 'static {
+    /// Insert the given ID into the set
+    fn insert(&self, id: usize);
+}
+
+// A collection of UnparkEvents to trigger on `unpark`
+#[derive(Clone)]
+struct Events {
+    set: Vec<UnparkEvent>, // TODO: change to some SmallVec
+}
+
+impl Events {
+    fn new() -> Events {
+        Events { set: Vec::new() }
+    }
+
+    fn trigger(&self) {
+        for event in self.set.iter() {
+            event.set.insert(event.item)
+        }
+    }
+
+    fn with_event(&self, event: UnparkEvent) -> Events {
+        let mut set = self.set.clone();
+        set.push(event);
+        Events { set: set }
     }
 }
 
@@ -420,96 +483,5 @@ impl Unpark for Inner {
             Ok(run) => self.exec.execute(run),
             Err(()) => {}
         }
-    }
-}
-
-/// For the duration of the given callback, add an "unpark event" to be
-/// triggered when the task handle is used to unpark the task.
-///
-/// Unpark events are used to pass information about what event caused a task to
-/// be unparked. In some cases, tasks are waiting on a large number of possible
-/// events, and need precise information about the wakeup to avoid extraneous
-/// polling.
-///
-/// Every `Task` handle comes with a set of unpark events which will fire when
-/// `unpark` is called. When fired, these events insert an identifer into a
-/// concurrent set, which the task can read from to determine what events
-/// occurred.
-///
-/// This function immediately invokes the closure, `f`, but arranges things so
-/// that `task::park` will produce a `Task` handle that includes the given
-/// unpark event.
-///
-/// # Panics
-///
-/// This function will panic if a task is not currently being executed. That
-/// is, this method can be dangerous to call outside of an implementation of
-/// `poll`.
-pub fn with_unpark_event<F, R>(event: UnparkEvent, f: F) -> R
-    where F: FnOnce() -> R
-{
-    with(|task, data| {
-        let new_task = Task {
-            id: task.id,
-            unpark: task.unpark.clone(),
-            events: task.events.with_event(event),
-        };
-        set(&new_task, data, f)
-    })
-}
-
-#[derive(Clone)]
-/// A set insertion to trigger upon `unpark`.
-///
-/// Unpark events are used to communicate information about *why* an unpark
-/// occured, in particular populating sets with event identifiers so that the
-/// unparked task can avoid extraneous polling. See `with_unpark_event` for
-/// more.
-pub struct UnparkEvent {
-    set: Arc<EventSet>,
-    item: usize,
-}
-
-impl UnparkEvent {
-    /// Construct an unpark event that will insert `id` into `set` when
-    /// triggered.
-    pub fn new(set: Arc<EventSet>, id: usize) -> UnparkEvent {
-        UnparkEvent {
-            set: set,
-            item: id,
-        }
-    }
-}
-
-/// A concurrent set which allows for the insertion of `usize` values.
-///
-/// `EventSet`s are used to communicate precise information about the event(s)
-/// that trigged a task notification. See `task::with_unpark_event` for details.
-pub trait EventSet: Send + Sync + 'static {
-    /// Insert the given ID into the set
-    fn insert(&self, id: usize);
-}
-
-// A collection of UnparkEvents to trigger on `unpark`
-#[derive(Clone)]
-struct Events {
-    set: Vec<UnparkEvent>, // TODO: change to some SmallVec
-}
-
-impl Events {
-    fn new() -> Events {
-        Events { set: Vec::new() }
-    }
-
-    fn trigger(&self) {
-        for event in self.set.iter() {
-            event.set.insert(event.item)
-        }
-    }
-
-    fn with_event(&self, event: UnparkEvent) -> Events {
-        let mut set = self.set.clone();
-        set.push(event);
-        Events { set: set }
     }
 }
